requirements.txt
Flask==0.12.2
Flask-RESTful==0.3.6
pynamodb==3.2.1
flask-restful-swagger-2==0.35
flask-swagger-ui==3.6.0
bcrypt==3.1.4
click==6.7
coverage==4.4.2
boto3==1.5.4
PyJWT==1.5.3
mongoengine==0.15.0
pymongo==3.6.0
mypy==0.560
Flask-Cors==3.0.3
arrow==0.12.0
PyFunctional==1.1.2
email-validator==1.0.3
